Change naming, add xml docs.
