+++ src/Client/Components/EntityTable/EntityClientTableContext.cs
@@ -1,13 +1,27 @@
 ﻿using FSH.BlazorWebAssembly.Client.Infrastructure.ApiClient;
 
-namespace FSH.BlazorWebAssembly.Client.Components.EntityManager;
+namespace FSH.BlazorWebAssembly.Client.Components.EntityTable;
 
-public class ClientEntityManagerContext<TEntity, TId> : EntityManagerContext<TEntity, TId>
+/// <summary>
+/// Initialization Context for the EntityTable Component.
+/// Use this one if you want to use Client Paging, Sorting and Filtering.
+/// </summary>
+/// <typeparam name="TEntity">The type of the entity.</typeparam>
+/// <typeparam name="TId">The type of the id of the entity.</typeparam>
+public class EntityClientTableContext<TEntity, TId> : EntityTableContext<TEntity, TId>
 {
+    /// <summary>
+    /// A function that loads all the data for the table from the api and returns a ListResult of TEntity.
+    /// </summary>
     public Func<Task<ListResult<TEntity>>> LoadDataFunc { get; }
+
+    /// <summary>
+    /// A function that returns a boolean which indicates whether the supplied entity meets the search criteria
+    /// (the supplied string is the search string entered).
+    /// </summary>
     public Func<string?, TEntity, bool> SearchFunc { get; }
 
-    public ClientEntityManagerContext(
+    public EntityClientTableContext(
         List<EntityField<TEntity>> fields,
         Func<Task<ListResult<TEntity>>> loadDataFunc,
         Func<string?, TEntity, bool> searchFunc,
