+++ src/Client/Components/EntityTable/EntityTable.razor
@@ -1,7 +1,7 @@
 ﻿@typeparam TEntity where TEntity: class, new()
 @typeparam TId where TId : IEquatable<TId>
 
-@inject IStringLocalizer<EntityManager<TEntity, TId>> L
+@inject IStringLocalizer<EntityTable<TEntity, TId>> L
 
 <style>
     .mud-table-container {
@@ -62,7 +62,7 @@
                     foreach (var field in Context.Fields)
                     {
                         <MudTh>
-                            @if (Context is ClientEntityManagerContext<TEntity, TId>)
+                            @if (Context is EntityClientTableContext<TEntity, TId>)
                             {
                                 <MudTableSortLabel SortBy="@field.DataValue">@field.DataLabel</MudTableSortLabel>
                             }
