+++ src/Client/Pages/Catalog/Products.razor
@@ -5,7 +5,7 @@
 
 <FshTitle Title="@L["Products"]" Description="@L["Manage Products."]" />
 
-<EntityManager TEntity="ProductDto" TId="Guid" Context="@Context">
+<EntityTable TEntity="ProductDto" TId="Guid" Context="@Context">
 
     <AdvancedSearchContent>
         <div style="margin-top:10px!important">
@@ -48,4 +48,4 @@
         </MudItem>
     </EditFormContent>
 
-</EntityManager>
+</EntityTable>
