+++ src/Client/Pages/Identity/Roles/Roles.razor
@@ -4,7 +4,7 @@
 
 <FshTitle Title="@L["Roles"]" Description="@L["Manage Roles."]" />
 
-<EntityManager TEntity="RoleDto" TId="string" Context="@Context">
+<EntityTable TEntity="RoleDto" TId="string" Context="@Context">
 
     <AdvancedSearchContent>
         <div style="margin-top:10px!important">
@@ -30,4 +30,4 @@
         </MudItem>
     </EditFormContent>
 
-</EntityManager>
+</EntityTable>
