+++ src/Client/_Imports.razor
@@ -25,7 +25,7 @@
 @using FSH.BlazorWebAssembly.Client.Pages.Authentication
 @using FSH.BlazorWebAssembly.Client.Pages.Multitenancy
 @using FSH.BlazorWebAssembly.Client.Shared.Components
-@using FSH.BlazorWebAssembly.Client.Components.EntityManager
+@using FSH.BlazorWebAssembly.Client.Components.EntityTable
 @using FSH.BlazorWebAssembly.Client.Components.ErrorHandler
 @using FSH.BlazorWebAssembly.Client.Components.Notifications
 @using FSH.BlazorWebAssembly.Client.Components.ThemeManager
